Update bigquery_data_access.view.lkml
removed the wildcard due to sharded table being replaced by partitioned. updated the where statement to use the partition field. This derived table is not needed anymore due to wildcard removed. This is will be removed in next patch (requires testing), prioritising this fix first as code is broken in current state.